[FIX] [Graph] 퍼센트가 아닌 실제 지출을 받아올 수 있도록 Long으로 변경

+++ app/src/main/java/com/woowacamp/android_accountbook_15/data/AccountBookDataSource.kt
@@ -42,7 +42,7 @@ interface AccountBookDataSource {
         startYear: Int, startMonth: Int,
         endYear: Int, endMonth: Int,
         categoryId: Long
-    ): List<Pair<Int, Int>>
+    ): List<Pair<Int, Long>>
 
     fun readAllPaymentMethod(): List<PaymentMethod>
